+++ src/index.html
@@ -27,7 +27,7 @@
 
   <footer class = "rodape">
     <a class="sobre" href="/#sobre">Sobre a Lemos</a>
-    <p class = "texto-rodapé">Desenvolvido por: Giselle Schwab, Melina Osik e Talita Martins, 2023.</p>
+    <p class = "texto-rodape">Desenvolvido por: Giselle Schwab, Melina Osik e Talita Martins, 2023.</p>
   </footer>
   <script src="index.js" type="module"></script>
 </body>
